improve(react-components): Empty 行を廃止して「セットを追加」ダイアログに統合 (#68)
#67 までの Empty 行 + Tabs+form は「追加と編集の動線」が異なる UI 形態を 2 種類
（Empty=ダイアログ、CreateSetPlanRow=プレビューボタン+パターントグル）持っていた。
design-principles.md L7 / L38 / L39 に照らすと:
- L7 「アプリの存在を忘れて計画に没入」: ボタン中身を読まなくても目的（追加）で押せる
- L38 「1,2,3 がわかれば 4 を予測できる」一貫性: 追加と編集で同じメンタルモデル
- L39 「大胆な取捨選択 / 目的達成への一直線」: 迂回 UI 撤廃、Tabs+form 1 経路へ
の方向で、追加 UI を Tabs+form ダイアログ 1 種類に統合する。

データモデル変更:
- SetPlan 型から { pattern: null } ケースを削除（params 必須に）
- #58 撤回: 初期セット 1 件生成は不要。種目計画は空 setPlans で生まれる

実装変更:
- 新規 SetPlanAddDialog: kind タブ + 値入力 + 確定 を 1 ダイアログで完結。
  直前セットがあればその kind と値を初期値として継承
- CreateSetPlanRow を SetPlanAddDialog の薄いラッパーに（プレビュー値ラベル /
  パターントグル / buildPayload / inferLastPattern を撤去）
- SetPlanRowEditTrigger の affordance variant のアイコンを PlusIcon に変更
  （edit variant の Pencil と区別）。aria-label は撤去し、accessible name は
  visible テキスト「セットを追加」自体になる
- SetPlanRowEmpty / その stories / build-payload / infer-last-pattern を削除
- SetPlanSection から Empty case 分岐 / hasEmptyRow / defaultPayloadFor を撤去、
  CreateSetPlanRow に weightStep を渡すように修正
- index.stories.tsx の AfterAddingExercisePlan を空 setPlans に更新
