+++ packages/react-components/src/views/program-detail/set-plan-row/index.ts
@@ -1,5 +1,4 @@
 export { CreateSetPlanRow } from "./create-set-plan-row";
-export { SetPlanRowEmpty } from "./set-plan-row-empty";
 export { SetPlanRowRepsXRpe } from "./set-plan-row-reps-x-rpe";
 export { SetPlanRowWeightXReps } from "./set-plan-row-weight-x-reps";
 export { SetPlanRowWeightXRpe } from "./set-plan-row-weight-x-rpe";
