+++ packages/react-components/src/views/program-detail/set-plan-types.ts
@@ -5,7 +5,7 @@ export type SetPlanWithParams =
 	| { pattern: "weight-x-rpe"; weight: number; rpe: number }
 	| { pattern: "reps-x-rpe"; reps: number; rpe: number };
 
-export type SetPlan = { id: string } & (SetPlanWithParams | { pattern: null });
+export type SetPlan = { id: string } & SetPlanWithParams;
 
 // 重量の単位。本来は Exercise の属性だが、現状 Exercise 型の抽出は不要なので
 // SetPlan 関連の型置き場である本ファイルに置く。将来 Exercise 型を分離する
